requirements.txt
pip==23.1.2
fastapi==0.98.0
uvicorn==0.22.0
tortoise-orm==0.19.3
asyncpg==0.28.0
argon2-cffi==21.3.0
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.1
password-strength==0.0.3.post2
